refactor(s3): Change offset types to uint64 and add configurable retry
- Change WriteAt and ReadAt offset parameter from int64 to uint64
  across all content store interfaces and implementations
- Remove unnecessary negative offset validation (now enforced by type)
- Add retry with exponential backoff for transient S3 errors:
  - Network errors, throttling (429), server errors (5xx)
  - Configurable via YAML: max_retries, initial_backoff, max_backoff,
    backoff_multiplier
  - Non-retryable errors (404, 403, invalid range) fail immediately
- Update NFS handlers to use uint64 offsets directly without casting
- Add retry configuration documentation to CONFIGURATION.md

+++ pkg/store/content/fs/fs_write.go
@@ -140,7 +140,7 @@ func (r *FSContentStore) WriteContent(ctx context.Context, id metadata.ContentID
 //
 // Returns:
 //   - error: Returns error if operation fails or context is cancelled
-func (r *FSContentStore) WriteAt(ctx context.Context, id metadata.ContentID, data []byte, offset int64) error {
+func (r *FSContentStore) WriteAt(ctx context.Context, id metadata.ContentID, data []byte, offset uint64) error {
 	// ========================================================================
 	// Step 1: Check context before filesystem operation
 	// ========================================================================
@@ -191,7 +191,7 @@ func (r *FSContentStore) WriteAt(ctx context.Context, id metadata.ContentID, dat
 		return err
 	}
 
-	_, err := file.Seek(offset, io.SeekStart)
+	_, err := file.Seek(int64(offset), io.SeekStart)
 	if err != nil {
 		return fmt.Errorf("failed to seek to offset: %w", err)
 	}
